ev3dev: document that LED use is racey when concurrent
It is possible to silence the race detector for access to the err field
and serialise device writes, but this does not address the issue that
call chains are not atomic.

Making call chains transactional adds complexity that does not seem
justified (the behaviour resulting from serialised concurrent access to
flashing lights does not seem to be meaningful in most cases). So leave
the memory race behaviour visible to be seen by the race detector, but
document the raciness.