ev3dev: remove Find as redundant
FindAfter(nil, ... works just as well. In fact we can implement all this
in the types them selves and use FindAfter as a wrapper (or delete it
too which would remove the need for reflect).

The scanning code is almost identical to deviceIDFor and these should
either be factored together or somehow else simplified.